+++ jobis-application/src/main/java/team/retum/jobis/domain/recruitment/usecase/ApplyRecruitmentUseCase.java
@@ -7,7 +7,6 @@ import team.retum.jobis.common.spi.SecurityPort;
 import team.retum.jobis.domain.company.model.Company;
 import team.retum.jobis.domain.recruitment.dto.request.ApplyRecruitmentRequest;
 import team.retum.jobis.domain.recruitment.event.InterestedRecruitmentEvent;
-import team.retum.jobis.domain.recruitment.event.WinterInternRegisteredEvent;
 import team.retum.jobis.domain.recruitment.exception.RecruitmentAlreadyExistsException;
 import team.retum.jobis.domain.recruitment.model.RecruitArea;
 import team.retum.jobis.domain.recruitment.model.Recruitment;
