+++ jobis-infrastructure/src/main/java/team/retum/jobis/event/intern/WinterInternEventHandler.java
@@ -6,10 +6,16 @@ import org.springframework.stereotype.Component;
 import org.springframework.transaction.event.TransactionPhase;
 import org.springframework.transaction.event.TransactionalEventListener;
 import team.retum.jobis.domain.auth.model.Authority;
+import team.retum.jobis.domain.company.exception.CompanyNotFoundException;
+import team.retum.jobis.domain.company.model.Company;
+import team.retum.jobis.domain.company.spi.QueryCompanyPort;
 import team.retum.jobis.domain.intern.event.WinterInternToggledEvent;
 import team.retum.jobis.domain.notification.model.Notification;
 import team.retum.jobis.domain.notification.model.Topic;
 import team.retum.jobis.domain.notification.spi.CommandNotificationPort;
+import team.retum.jobis.domain.intern.event.WinterInternRegisteredEvent;
+import team.retum.jobis.domain.recruitment.model.RecruitStatus;
+import team.retum.jobis.domain.recruitment.model.Recruitment;
 import team.retum.jobis.domain.user.model.User;
 import team.retum.jobis.domain.user.spi.QueryUserPort;
 import team.retum.jobis.thirdparty.fcm.FCMUtil;
@@ -23,6 +29,7 @@ public class WinterInternEventHandler {
     private final FCMUtil fcmUtil;
     private final CommandNotificationPort commandNotificationPort;
     private final QueryUserPort queryUserPort;
+    private final QueryCompanyPort queryCompanyPort;
 
     @Async("asyncTaskExecutor")
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
@@ -48,5 +55,39 @@ public class WinterInternEventHandler {
                 fcmUtil.sendMessageToTopic(notification);
             }
         });
+    }
+
+    @Async("asyncTaskExecutor")
+    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
+    public void onWinterInternRegistered(WinterInternRegisteredEvent event) {
+        List<Recruitment> recruitments = event.getRecruitments();
+
+        for (Recruitment recruitment : recruitments) {
+            if (recruitment.isWinterIntern() && recruitment.getStatus() == RecruitStatus.RECRUITING) {
+                List<String> deviceTokens = queryUserPort.getDeviceTokenByTopic(Topic.WINTER_INTERN);
+
+                Company company = queryCompanyPort.getById(recruitment.getCompanyId())
+                    .orElseThrow(() -> CompanyNotFoundException.EXCEPTION);
+
+                String companyName = company.getName();
+
+                deviceTokens.forEach(deviceToken -> {
+                    User user = queryUserPort.getUserIdByDeviceToken(deviceToken);
+
+                    Notification notification = Notification.builder()
+                        .title(companyName + " 겨울 인턴십 모집 공고 ⛄️")
+                        .content("겨울 인턴십 모집 의뢰서가 등록되었어요. 지금 확인해보세요!")
+                        .userId(user.getId())
+                        .detailId(recruitment.getId())
+                        .topic(Topic.WINTER_INTERN)
+                        .authority(Authority.STUDENT)
+                        .isNew(true)
+                        .build();
+
+                    commandNotificationPort.save(notification);
+                    fcmUtil.sendMessageToTopic(notification);
+                });
+            }
+        }
     }
 }
