+++ jobis-infrastructure/src/main/java/team/retum/jobis/event/recruitment/RecruitmentEventHandler.java
@@ -8,16 +8,12 @@ import org.springframework.transaction.event.TransactionalEventListener;
 import team.retum.jobis.domain.auth.model.Authority;
 import team.retum.jobis.domain.bookmark.spi.QueryBookmarkPort;
 import team.retum.jobis.domain.bookmark.spi.vo.BookmarkUserVO;
-import team.retum.jobis.domain.company.exception.CompanyNotFoundException;
-import team.retum.jobis.domain.company.model.Company;
 import team.retum.jobis.domain.company.spi.QueryCompanyPort;
 import team.retum.jobis.domain.notification.model.Notification;
 import team.retum.jobis.domain.notification.model.Topic;
 import team.retum.jobis.domain.notification.spi.CommandNotificationPort;
 import team.retum.jobis.domain.recruitment.event.InterestedRecruitmentEvent;
 import team.retum.jobis.domain.recruitment.event.RecruitmentStatusChangedEvent;
-import team.retum.jobis.domain.recruitment.event.WinterInternRegisteredEvent;
-import team.retum.jobis.domain.recruitment.model.RecruitStatus;
 import team.retum.jobis.domain.recruitment.model.Recruitment;
 import team.retum.jobis.domain.recruitment.spi.QueryRecruitAreaPort;
 import team.retum.jobis.domain.student.model.Student;
@@ -109,40 +105,6 @@ public class RecruitmentEventHandler {
             }
 
             fcmUtil.sendMessages(repNotification, tokens);
-        }
-    }
-
-    @Async("asyncTaskExecutor")
-    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
-    public void onWinterInternRegistered(WinterInternRegisteredEvent event) {
-        List<Recruitment> recruitments = event.getRecruitments();
-
-        for (Recruitment recruitment : recruitments) {
-            if (recruitment.isWinterIntern() && recruitment.getStatus() == RecruitStatus.RECRUITING) {
-                List<String> deviceTokens = queryUserPort.getDeviceTokenByTopic(Topic.WINTER_INTERN);
-
-                Company company = queryCompanyPort.getById(recruitment.getCompanyId())
-                    .orElseThrow(() -> CompanyNotFoundException.EXCEPTION);
-
-                String companyName = company.getName();
-
-                deviceTokens.forEach(deviceToken -> {
-                    User user = queryUserPort.getUserIdByDeviceToken(deviceToken);
-
-                    Notification notification = Notification.builder()
-                        .title(companyName + " 겨울 인턴십 모집 공고 ⛄️")
-                        .content("겨울 인턴십 모집 의뢰서가 등록되었어요. 지금 확인해보세요!")
-                        .userId(user.getId())
-                        .detailId(recruitment.getId())
-                        .topic(Topic.WINTER_INTERN)
-                        .authority(Authority.STUDENT)
-                        .isNew(true)
-                        .build();
-
-                    commandNotificationPort.save(notification);
-                    fcmUtil.sendMessageToTopic(notification);
-                });
-            }
         }
     }
 }
